Split yaml properties completion contributor into key/value contributor
