+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/spring/properties/SpringYamlQuickCompletionContributor.java
@@ -1,35 +1,44 @@
 package com.microsoft.azure.toolkit.intellij.connector.spring.properties;
 
 import com.intellij.codeInsight.completion.*;
+import com.intellij.openapi.project.DumbAware;
 import com.intellij.patterns.PsiJavaPatterns;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.ProcessingContext;
 import com.microsoft.azure.toolkit.intellij.connector.completion.LookupElements;
 import com.microsoft.azure.toolkit.intellij.connector.spring.SpringSupported;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
+import org.jetbrains.yaml.YAMLUtil;
 import org.jetbrains.yaml.psi.YAMLKeyValue;
+import org.jetbrains.yaml.psi.YAMLPsiElement;
 
 import javax.annotation.Nonnull;
 import java.util.List;
+import java.util.Objects;
 
-public class SpringYamlPropertiesQuickCompletionContributor extends CompletionContributor {
+public class SpringYamlQuickCompletionContributor extends CompletionContributor implements DumbAware {
 
     public static final String APPLICATION = "application";
 
-    public SpringYamlPropertiesQuickCompletionContributor() {
+    public SpringYamlQuickCompletionContributor() {
         super();
-        extend(CompletionType.BASIC, SpringYamlPropertiesCompletionContributor.YAML_TEXT.withSuperParent(2, PsiJavaPatterns.psiElement(YAMLKeyValue.class)), new CompletionProvider<>() {
+        extend(CompletionType.BASIC, SpringYamlCompletionContributor.YAML_TEXT.withSuperParent(2, PsiJavaPatterns.psiElement(YAMLKeyValue.class)), new CompletionProvider<>() {
             @Override
             protected void addCompletions(@Nonnull final CompletionParameters parameters, @Nonnull final ProcessingContext context, @Nonnull final CompletionResultSet result) {
                 final PsiElement element = parameters.getPosition();
-                final String key = SpringYamlPropertiesCompletionProvider.getYamlElementKey(element);
+                final YAMLPsiElement yamlElement = PsiTreeUtil.getParentOfType(element, YAMLPsiElement.class);
+                if (Objects.isNull(yamlElement)) {
+                    return;
+                }
+                final String key = YAMLUtil.getConfigFullName(yamlElement);
                 final List<? extends SpringSupported<?>> definitions = SpringPropertyValueCompletionProvider.getSupportedDefinitions(key);
                 if (!definitions.isEmpty()) {
                     if (!Azure.az(AzureAccount.class).isLoggedIn()) {
                         result.addElement(LookupElements.buildSignInLookupElement());
                     } else {
-                        result.addElement(LookupElements.buildConnectLookupElement(definitions.get(0), SpringPropertyValueCompletionProvider.PropertyValueInsertHandler::insert));
+                        result.addElement(LookupElements.buildConnectLookupElement(definitions.get(0), YamlUtils::insertYamlConnection));
                     }
                 }
             }
